I'll have to get back to gltf later

+++ Chapter 6 ECS/Components.h
@@ -20,7 +20,7 @@ struct Texture {
 
 struct Mesh {
 	unsigned int arrayID;
-	unsigned int numVerts;
+	size_t numVerts;
 };
 
 struct VertexArray {
